pyproject.toml
[tool.poetry]
name = "socialapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.0"}
ruff = "^0.6.7"
pytest = "^8.3.3"
sqlalchemy = "^2.0.35"
databases = {extras = ["aiosqlite"], version = "^0.9.0"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
rich = "^13.9.2"
asgi-correlation-id = "^4.3.3"
python-json-logger = "^2.0.7"
python-jose = "^3.3.0"
python-multipart = "^0.0.12"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
